requirements.txt
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
clearbit==0.1.7
click==7.1.2
cryptography==3.2.1
dominate==2.6.0
Flask==1.1.2
Flask-Bootstraps==3.3.7.8
Flask-JWT-Extended==3.24.1
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
pyhunter==1.7
PyJWT==1.7.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.25.11
visitor==0.1.3
Werkzeug==1.0.1
